+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/SentryBridge.apple.kt
@@ -65,6 +65,10 @@ internal actual object SentryBridge {
         SentrySDK.setUser(user?.toCocoaUser())
     }
 
+    actual fun isCrashedLastRun(): Boolean {
+        return SentrySDK.crashedLastRun()
+    }
+
     actual fun close() {
         SentrySDK.close()
     }